rename resources and docstrings
