+++ setup.py
@@ -1,4 +1,4 @@
-"""Setup TiVTiler."""
+"""Setup timvt."""
 
 from setuptools import find_packages, setup
 
@@ -7,11 +7,9 @@ with open("README.md") as f:
 
 inst_reqs = [
     "buildpg",
-    "fastapi",
-    "jinja2",
+    "fastapi[full]",
     "asyncpg",
-    "morecantile~=2.0.1",
-    "email-validator",
+    "morecantile>=2.1,<2.2",
     "dataclasses;python_version<'3.7'",
 ]
 extra_reqs = {
@@ -25,7 +23,6 @@ extra_reqs = {
         "mapbox-vector-tile",
     ],
     "dev": ["pytest", "pytest-cov", "pytest-asyncio", "requests", "pre-commit"],
-    "server": ["uvicorn"],
     "docs": ["nbconvert", "mkdocs", "mkdocs-material", "mkdocs-jupyter", "pygments"],
 }
 
@@ -36,7 +33,7 @@ setup(
     description=u"",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    python_requires=">=3",
+    python_requires=">=3.6",
     classifiers=[
         "Intended Audience :: Information Technology",
         "Intended Audience :: Science/Research",
