+++ timvt/custom/__init__.py
@@ -1,1 +1,1 @@
-"""TiVTiler.custom"""
+"""timvt.custom"""
