+++ timvt/db/__init__.py
@@ -1,1 +1,1 @@
-"""TiVTiler.db"""
+"""timvt.db"""
