+++ timvt/endpoints/__init__.py
@@ -1,1 +1,1 @@
-"""TiVTiler.endpoints."""
+"""timvt.endpoints."""
