+++ timvt/endpoints/factory.py
@@ -1,4 +1,4 @@
-"""TiVTiler.endpoints.factory: router factories."""
+"""timvt.endpoints.factory: router factories."""
 
 from dataclasses import dataclass, field
 from typing import Any, Callable, Dict, List, Optional, Type
@@ -10,7 +10,7 @@ from timvt.db.tiles import VectorTileReader
 from timvt.dependencies import TableParams, TileMatrixSetParams, _get_db_pool
 from timvt.models.mapbox import TileJSON
 from timvt.models.metadata import TableMetadata
-from timvt.ressources.enums import MimeTypes
+from timvt.resources.enums import MimeTypes
 from timvt.utils import Timer
 
 from fastapi import APIRouter, Depends, Path, Query
