+++ timvt/models/__init__.py
@@ -1,1 +1,1 @@
-"""TiVTiler.models"""
+"""timvt.models"""
