--- timvt/ressources/__init__.py
@@ -1,1 +1,0 @@
-"""TiVTiler.ressources."""
